feat: FallingAmountView - 임시 반영
