+++ app/src/main/res/drawable/bg_center_transparent_gradient.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:type="linear"
+        android:angle="270"
+        android:startColor="#FFFFFF"
+        android:centerColor="#00FFFFFF"
+        android:endColor="#FFFFFF" />
+</shape>
